+++ app/api/scraper/controller/scraper_controller.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, HTTPException
 from fastapi.responses import JSONResponse
-from app.dto.requests.process_request_dto import ProcessRequest
+from app.dto.request.process_request import ProcessRequest
 import subprocess
 import os
 import json
